마이페이지 회원정보 수정 완성

+++ backend/skip/src/main/java/com/example/skip/controller/UserSocialController.java
@@ -6,12 +6,16 @@ import com.example.skip.enumeration.UserSocial;
 import com.example.skip.service.UserSocialService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.io.IOException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 
 @Controller
 @RequestMapping("/user/social")
@@ -30,22 +34,49 @@ public class UserSocialController {
     }
 
     @GetMapping("/link/{client}")
-    public void link(@PathVariable("client") String client,
-                     @RequestParam("code") String authorizationCode,
-                     @AuthenticationPrincipal UserDetails userDetails,
-                     HttpServletResponse response) throws IOException {
+    public String link(@PathVariable("client") String client,
+                       @RequestParam("code") String authorizationCode,
+                       @AuthenticationPrincipal UserDetails userDetails,
+                       RedirectAttributes redirectAttributes) throws IOException {
 
         UserDto userDto = (UserDto) userDetails;
 
-        userSocialService.link(UserSocial.valueOf(client.toUpperCase()), authorizationCode, userDto);
+        Boolean success;
+        String data = "";
 
-        response.sendRedirect("http://localhost:5173/mypage/account");
+        try {
+            userSocialService.link(UserSocial.valueOf(client.toUpperCase()), authorizationCode, userDto);
+            success = true;
+        } catch (DataIntegrityViolationException e) {
+            e.printStackTrace();
+            success = false;
+            data = URLEncoder.encode("이미 연동된 소셜 계정입니다.", StandardCharsets.UTF_8);
+        }
+
+        return "redirect:http://localhost:5173/mypage/account?success=" + success + "&data=" + data;
     }
 
     @ResponseBody
     @DeleteMapping("/unlink")
     public ApiResponse unlink(@AuthenticationPrincipal UserDetails userDetails) {
         UserDto userDto = (UserDto) userDetails;
+
+        Boolean unset = false;
+        if (userDto.getSocial() == UserSocial.KAKAO) {
+            if (!userDto.getKakaoLinkageDto().getUsernameSet() || !userDto.getKakaoLinkageDto().getPasswordSet()) {
+                unset = true;
+            }
+        } else if (userDto.getSocial() == UserSocial.NAVER) {
+            if (!userDto.getNaverLinkageDto().getUsernameSet() || !userDto.getNaverLinkageDto().getPasswordSet()) {
+                unset = true;
+            }
+        }
+        if (unset) {
+            return ApiResponse.builder()
+                    .success(false)
+                    .data("아이디, 비밀번호를 먼저 설정해주세요.").build();
+        }
+
         try {
             String accessToken = userSocialService.getAccessToken(userDto);
             userSocialService.unlink(userDto, accessToken);
